Chatbot/requirements.txt
streamlit
langchain
langchain-community
huggingface_hub
python-dotenv
